% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_match_helpers.R
\name{.attach_metadata_peru}
\alias{.attach_metadata_peru}
\title{Attach Metadata to Results}
\usage{
.attach_metadata_peru(tbl, n_input, n_matched, n_fuzzy_genus, n_fuzzy_species)
}
\description{
Attach Metadata to Results
}
\keyword{internal}
