% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_match_helpers.R
\name{.combine_unmatched_nodes_peru}
\alias{.combine_unmatched_nodes_peru}
\title{Combine Unmatched Nodes}
\usage{
.combine_unmatched_nodes_peru(pipe, invalid_df)
}
\description{
Combine Unmatched Nodes
}
\keyword{internal}
