% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_match_helpers.R
\name{.compute_matched_rank_peru}
\alias{.compute_matched_rank_peru}
\title{Compute Matched Rank for Peru Mammals}
\usage{
.compute_matched_rank_peru(df)
}
\description{
Compute Matched Rank for Peru Mammals
}
\keyword{internal}
