% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_match_helpers.R
\name{.consolidate_ambiguous_attrs_peru}
\alias{.consolidate_ambiguous_attrs_peru}
\title{Consolidate Ambiguous Match Attributes}
\usage{
.consolidate_ambiguous_attrs_peru(output, pipe)
}
\description{
Consolidate Ambiguous Match Attributes
}
\keyword{internal}
