% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.detect_trinomial}
\alias{.detect_trinomial}
\title{Detect trinomial names (3+ taxonomic elements)}
\usage{
.detect_trinomial(orig_names)
}
\description{
Detect trinomial names (3+ taxonomic elements)
}
\keyword{internal}
