% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_match_helpers.R
\name{.final_assertions_peru}
\alias{.final_assertions_peru}
\title{Final Validation of Results}
\usage{
.final_assertions_peru(splist_class, output)
}
\description{
Final Validation of Results
}
\keyword{internal}
