% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_match_helpers.R
\name{.format_matched_names_peru}
\alias{.format_matched_names_peru}
\title{Format Matched Names for Display}
\usage{
.format_matched_names_peru(df)
}
\description{
Format Matched Names for Display
}
\keyword{internal}
