% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_match_helpers.R
\name{.init_matching_columns_peru}
\alias{.init_matching_columns_peru}
\title{Initialize Matching Columns}
\usage{
.init_matching_columns_peru(df)
}
\description{
Initialize Matching Columns
}
\keyword{internal}
