% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.invalidate_trinomials}
\alias{.invalidate_trinomials}
\title{Invalidate trinomial matches in validation results}
\usage{
.invalidate_trinomials(results)
}
\description{
Invalidate trinomial matches in validation results
}
\keyword{internal}
