% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.onAttach}
\alias{.onAttach}
\title{.onAttach hook}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{A character string indicating the path to the library.}

\item{pkgname}{A character string with the name of the package.}
}
\description{
Hook function that runs when the package is attached via \code{library()}.
It displays the package version and information about the taxonomic backbone.
}
\keyword{internal}
