% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.onLoad}
\alias{.onLoad}
\title{.onLoad hook}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{A character string with the name of the library directory.}

\item{pkgname}{A character string with the name of the package.}
}
\description{
Hook function that runs when the package is loaded.
It sets default options for the package.
}
\keyword{internal}
