% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_match_helpers.R
\name{.split_valid_invalid_peru}
\alias{.split_valid_invalid_peru}
\title{Split Valid and Invalid Names}
\usage{
.split_valid_invalid_peru(splist_class)
}
\description{
Split Valid and Invalid Names
}
\keyword{internal}
