% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_match_helpers.R
\name{.validate_target_schema_peru}
\alias{.validate_target_schema_peru}
\title{Validate Target Database Schema}
\usage{
.validate_target_schema_peru(target_df)
}
\description{
Validate Target Database Schema
}
\keyword{internal}
