% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perumammals-data.R
\docType{data}
\name{peru_mammals_backbone}
\alias{peru_mammals_backbone}
\title{Summary information on the \pkg{perumammals} taxonomic backbone}
\format{
A tibble with 1 row and 4 variables:
\describe{
\item{source}{Character. Short bibliographic reference to the backbone
source (Pacheco et al. 2021).}
\item{source_year}{Integer. Publication year of the backbone source
(2021).}
\item{n_species}{Integer. Number of species included in the backbone
(as rows in \code{\link{peru_mammals}}).}
\item{created_at}{Date. Date when the backbone data objects were
generated (in the package build process).}
}
}
\usage{
data("peru_mammals_backbone")
}
\description{
A one-row tibble with metadata about the taxonomic backbone used in
\pkg{perumammals}, including its bibliographic source, year, number
of species and the date when the internal data objects were created.
}
\details{
This object is intended for internal bookkeeping and for functions that
report the origin and version of the backbone.
}
\seealso{
\code{\link{peru_mammals}}
}
\keyword{datasets}
