% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{show_progress}
\alias{show_progress}
\title{Determine whether to show progress bar}
\usage{
show_progress()
}
\value{
Logical indicating whether progress bars should be shown.
}
\description{
Returns logical TRUE/FALSE depending on package options and whether
the session is interactive.
}
\keyword{internal}
